Enemy Detection
Added the definition of the enemy, now the skeletons do not hit each other. The enemy is defined in the EnemyDetector script, which hangs on the object. You can add an object and its enemies to the list (In this case, the enemies do not need to add it, because adding works like adding vertices to an undirected graph).